requirements.txt
streamlit
opencv-python-headless
scenedetect
Pillow